Include debug information in the generated module
This commit ensures that the name section and debug info is kept in the module generated by running `javy compile <source>.wasm`, this makes it easier to profile QuickJS and any APIs implemented in Rust. 

Note: this behaviour should be made configurable if we want to port this change upstream.