Handle updated bytecode format
This commit updates all the existing crates in order to handle and
updated QuickJS bytecode format, introduced in version 0.10.0 of
`rquickjs`.

Notably the main difference is the position of the constant pool for
each function. Prior to this update the code assumed that the constant
pool would come last, per function. As of version 0.10.0 of rquickjs,
the bytecode comes before the oeprators section.

This commit also registers the crates in the Makefile targets for
testing.