Add the compiler runtime plugin
This commit adds the compiler runtime to Javy as a WASI Preview 1
module.

The functionality is still a work-in-progress as of this commit.

The objective is to first work towards fully functioning calls between
Wasm and the runtime, which involves figuring out what's the best way
to handle closure creation as well as the function frame setup.