Initial runtime consumer
Got an initial (but incomplete) version of the consumer working.
Incomplete since it only handles Wasm -> QuickJS calls, the important
thing is that it respects object lifetimes et al, so we can keep a
trace of the objects and so useful things with them, like calling a
function.

The next iteration will involve creating the QuickJS -> Wasm piece,
via funcrefs.