Progress on lowering infrastructure: function prelude
This commit is a checkpoint for the compiler Wasm lowering
infrastructure, more concretely this commit handles the function's
prelude by declaring the locals, handling the function's entry and
exit blocks.

Additionally this commits adds the first couple of methods to the
abstract stack, which will be used when handling each operator. As a
matter of exemplification, this commit introduces the handling of the
`Drop` operator.