Allow standalone function definitions
Prior to this commit, the notion of a `FunctionBuilder` was tied to a
the frontend, that is, the code responsible for translating QuickJS
bytecode to SSA. With the introduction function calls, in which
we need to manually define trampolines for functions reachable from
QuickJS, there's a need for a mechanism that allows defining functions
directly through a builder.

This commit introduces a refactor in the compiler to meet such
requirement, enabling a better separation of concerns:

* The frontend is resposible for translating the bytecode, with help
from the `FunctionBuilder`, but they are not tied.
* The `FunctionBuilder` can now be used independently to generate
function bodies.

The next step is to fill in the `define_qjs_to_wasm_trampoline` function.