Improve compiler runtime builtin handling
This commit improves the way compiler runtime builtins are
handled. Concretely, it formally defines each builtin as
`RuntimeFunction` in the `crt` module.

Additionally, to avoid at this stage, doing a preliminary pass to
collect all the function imports needed, this change eagerly imports
all the compiler runtime functions at the start of the
compilation. This approach can obviously be improved by doing a
preliminary pass of the bytecode to figure out which imports are
actually needed. This is left as an exercise for later on, given that
there are other means to optimize dead imports e.g., `waffle` might
already do this or via `wasm-opt`

Lastly, this commit hooks the compiler to javy's frontend, allowing
the generation of code in an automated way.